feat: Modify 'handleDelte' function in 'ToDoList' component.
Add a new 'reset' action in the 'todos-reducer' to move deleted 'todolist' items
to the 'todontlist' instead of deleting them completely.
Replace the 'handleDelete' function to the 'handleReset' and dispath the 'reset' action.
Pass the 'handleReset' function to the 'handleReset' props within the 'List' and 'Item' components.
Add a new 'reset' button in the 'Item' component.