feat: Auto move 'todontlist' to 'todolist'.
Add a condition statement whithin the 'handleCheck' and 'handleRest' funciont in the 'ToDolist' component
to check the lenth of 'todolist' item is over three,and if so, return without dispatching the 'checkedStart' action.
Otherwise, dispatch the 'chckedStart' action to automatically move the 'todontlist' item to the 'todolist'
when one 'todolist' item is checked complete.

Refactor the 'checkedStart' action by removing the 'checked' property from the action since the value of 'startYn' always true.