feat: Mark rollback of reseting 'donelist' item
Refactor the 'resetDone' action by removing the 'doneDate' property that is set null
to add a '[rollback]' mark to each reset 'todolist' item via the 'doneDate' property.
Update the 'Item' component to display the '[rollback]' mark
if the value of 'startYn' is true and the value of 'doneDate' is not null.