feat: Add the 'darkMode'
Create a new 'DarkModeContext' in the app.js and set 'darkMode' state using the 'useState' hook.
Provide  the 'DarkModeContext' in the 'App.js'
Consume the 'DarkModeContext' in the 'NavBar.jsx'
Add a new 'dark' or 'red' button to the 'NavBar' component
and add an 'onClick' function to change theme.
The function shoult toggle the value of 'darkMode' useing the 'setToggle' function.