feat: Add a new 'sorted' action
Add a new 'priority' property to the each 'todos' item.
Add a new 'sorted' action in the todos-reducer to handle reordering of todos.
Add the new 'onEnd' event to the 'ReactSortable' component,
whitch triggers when a list item is dragged to a new position.
Disptch the 'sorted' action update the 'todos' state with the new order of items.