Refactor: Refactor the 'sorted' action.
Modify the 'getFilterdList' utility function by adding the 'sort' method with the 'compareIndex'
to avoid using the 'sord' method repeatedly.
Remove the 'sort' method from the 'todonlist' variable in the 'ToDontList' component.
Refactor the 'sorted' action by Adding a new 'filterName' argunment and refactoring logic,
so that the action can be used  in the 'ToDoList' component.
Add the 'sorted' action to the 'ToDoList' component.